fix: move google tag manager script to _app
